Correct query, correct statement executions

+++ src/main/java/io/supertokens/storage/postgresql/queries/DeleteUserQuery.java
@@ -66,7 +66,7 @@ public class DeleteUserQuery {
 
         try (PreparedStatement pst = conn.prepareStatement(query)) {
             pst.setString(1, userId);
-            pst.executeQuery();
+            pst.executeUpdate();
         }
     }
 
@@ -76,7 +76,7 @@ public class DeleteUserQuery {
         try (PreparedStatement pst = conn.prepareStatement(query)) {
 
             pst.setString(1, userId);
-            pst.executeQuery();
+            pst.executeUpdate();
         }
     }
 
@@ -85,7 +85,7 @@ public class DeleteUserQuery {
 
         try (PreparedStatement pst = conn.prepareStatement(query)) {
             pst.setString(1, userId);
-            pst.executeQuery();
+            pst.executeUpdate();
         }
     }
 
@@ -101,7 +101,7 @@ public class DeleteUserQuery {
     }
 
     private String[] getUserSessionHandles(Connection conn, @Nonnull String userId) throws SQLException {
-        String query = "SELECT user_id FROM " + Config.getConfig(start).getSessionInfoTable() + " WHERE user_id = ?";
+        String query = "SELECT session_handle FROM " + Config.getConfig(start).getSessionInfoTable() + " WHERE user_id = ?";
 
         try (PreparedStatement pst = conn.prepareStatement(query)) {
             pst.setString(1, userId);
@@ -120,7 +120,9 @@ public class DeleteUserQuery {
     private void deleteUserSessionHandles(Connection conn, @Nonnull String userId) throws SQLException {
         String[] sessionHandles = getUserSessionHandles(conn, userId);
 
-        System.out.println(String.join(", ", sessionHandles));
+        if (sessionHandles.length == 0) {
+            return;
+        }
 
         StringBuilder queryBuilder = new StringBuilder("DELETE FROM " + Config.getConfig(start).getSessionInfoTable() + " WHERE session_handle IN ");
 
@@ -142,7 +144,7 @@ public class DeleteUserQuery {
                 pst.setString(pstIndex, sessionHandles[i]);
             }
 
-            pst.executeQuery();
+            pst.executeUpdate();
         }
     }
 
@@ -152,7 +154,7 @@ public class DeleteUserQuery {
 
         try (PreparedStatement pst = conn.prepareStatement(query)) {
             pst.setString(1, userId);
-            pst.executeQuery();
+            pst.executeUpdate();
         }
     }
 
@@ -161,7 +163,7 @@ public class DeleteUserQuery {
 
         try (PreparedStatement pst = conn.prepareStatement(query)) {
             pst.setString(1, userId);
-            pst.executeQuery();
+            pst.executeUpdate();
         }
     }
 }
